data-pipeline/requirements.txt
gdal==3.8.0
rasterio==1.3.9
numpy==1.26.2
requests==2.31.0
boto3==1.34.10
psycopg2-binary==2.9.9
pandas==2.1.4
geopandas==0.14.1
click==8.1.7
tqdm==4.66.1
Pillow==10.1.0
